+++ Engine/Core/RHI/ShaderType.cpp
@@ -1,0 +1,5 @@
+// ShaderType.cpp.cc
+// created on 2021/6/6
+// author @zoloypzuo
+#include "ZeloPreCompiledHeader.h"
+#include "ShaderType.h"
