+++ Engine/Renderer/OpenGL/GLVertexArray.cpp
@@ -3,40 +3,10 @@
 // author @zoloypzuo
 #include "ZeloPreCompiledHeader.h"
 #include "GLVertexArray.h"
+#include "Core/RHI/ShaderType.h"
+#include "Renderer/OpenGL/GLUtil.h"
 
 namespace Zelo {
-
-static GLenum ShaderDataTypeToOpenGLBaseType(ShaderDataType type) {
-    switch (type) {
-        case ShaderDataType::Float:
-            return GL_FLOAT;
-        case ShaderDataType::Float2:
-            return GL_FLOAT;
-        case ShaderDataType::Float3:
-            return GL_FLOAT;
-        case ShaderDataType::Float4:
-            return GL_FLOAT;
-        case ShaderDataType::Mat3:
-            return GL_FLOAT;
-        case ShaderDataType::Mat4:
-            return GL_FLOAT;
-        case ShaderDataType::Int:
-            return GL_INT;
-        case ShaderDataType::Int2:
-            return GL_INT;
-        case ShaderDataType::Int3:
-            return GL_INT;
-        case ShaderDataType::Int4:
-            return GL_INT;
-        case ShaderDataType::Bool:
-            return GL_BOOL;
-        default:
-            break;
-    }
-
-    ZELO_CORE_ASSERT(false, "Unknown ShaderDataType!");
-    return 0;
-}
 
 GLVertexArray::GLVertexArray() {
     ZELO_PROFILE_FUNCTION();
